change type

+++ Final Project/Common/src/main/java/commonproject/model/menu/Menu.java
@@ -26,5 +26,5 @@ public class Menu {
     private FoodCategory foodCategory;
 
     @NotNull
-    private float pricePerItem;
+    private BigDecimal pricePerItem;
 }
